Implementd AccordionUIController
